pyproject.toml
[project]
name = "newspulse"
version = "0.1.0"
description = "Telegram bot that monitors news sources for user-defined topics"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "feedparser>=6.0",
    "aiosqlite>=0.20",
    "apscheduler>=3.10,<4",
    "pydantic-settings>=2.0",
    "lxml>=5.0",
    "google-genai>=1.68.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "ruff>=0.6"]

[project.scripts]
newspulse = "newspulse.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/newspulse"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
]
